Moved DimOpLoweringPass to EarlyLowering.cpp
